Split sockets endpoints into one file each, shared helpers in base.go
